move-files-and-2-renames (#4285)
